refactor(did): extract shared Ed25519 key utilities to ed25519Utils.mts
